Replace quasiharmonic tests
